requirements.txt
Django==5.0.7
djangorestframework==3.15.2
PyJWT==2.8.0